Extract identifiers to file
