+++ packages/app/src/cli/models/app/identifiers.ts
@@ -1,0 +1,103 @@
+import {Extension} from './extensions.js'
+import {App} from './app'
+import {dotEnvFileNames} from '../../constants.js'
+import {path, string} from '@shopify/cli-kit'
+import {writeDotEnv} from '@shopify/cli-kit/node/dot-env'
+
+export interface IdentifiersExtensions {
+  [localIdentifier: string]: string
+}
+
+export interface Identifiers {
+  /** Application's API Key */
+  app: string
+
+  /**
+   * The extensions' unique identifiers.
+   */
+  extensions: IdentifiersExtensions
+
+  /**
+   * The extensions' numeric identifiers (expressed as a string).
+   */
+  extensionIds: IdentifiersExtensions
+}
+
+export type UuidOnlyIdentifiers = Omit<Identifiers, 'extensionIds'>
+type UpdateAppIdentifiersCommand = 'dev' | 'deploy'
+interface UpdateAppIdentifiersOptions {
+  app: App
+  identifiers: UuidOnlyIdentifiers
+  command: UpdateAppIdentifiersCommand
+}
+/**
+ * Given an app and a set of identifiers, it persists the identifiers in the .env files.
+ * @param options {UpdateAppIdentifiersOptions} Options.
+ * @returns {App} An copy of the app with the environment updated to reflect the updated identifiers.
+ */
+
+export async function updateAppIdentifiers(
+  {app, identifiers, command}: UpdateAppIdentifiersOptions,
+  systemEnvironment = process.env,
+): Promise<App> {
+  let dotenvFile = app.dotenv
+  if (!dotenvFile) {
+    dotenvFile = {
+      path: path.join(app.directory, dotEnvFileNames.production),
+      variables: {},
+    }
+  }
+  const updatedVariables: {[key: string]: string} = {...(app.dotenv?.variables ?? {})}
+  if (!systemEnvironment[app.idEnvironmentVariableName]) {
+    updatedVariables[app.idEnvironmentVariableName] = identifiers.app
+  }
+  Object.keys(identifiers.extensions).forEach((identifier) => {
+    const envVariable = `SHOPIFY_${string.constantize(identifier)}_ID`
+    if (!systemEnvironment[envVariable]) {
+      updatedVariables[envVariable] = identifiers.extensions[identifier]
+    }
+  })
+
+  const write = JSON.stringify(dotenvFile.variables) !== JSON.stringify(updatedVariables) && command === 'deploy'
+  dotenvFile.variables = updatedVariables
+  if (write) {
+    await writeDotEnv(dotenvFile)
+  }
+  return {
+    ...app,
+    dotenv: dotenvFile,
+  }
+}
+interface GetAppIdentifiersOptions {
+  app: App
+}
+/**
+ * Given an app and a environment, it fetches the ids from the environment
+ * and returns them.
+ * @param options {GetAppIdentifiersOptions} Options.
+ * @returns
+ */
+
+export function getAppIdentifiers(
+  {app}: GetAppIdentifiersOptions,
+  systemEnvironment = process.env,
+): Partial<UuidOnlyIdentifiers> {
+  const envVariables = {
+    ...app.dotenv?.variables,
+    ...(systemEnvironment as {[variable: string]: string}),
+  }
+  const extensionsIdentifiers: {[key: string]: string} = {}
+  const processExtension = (extension: Extension) => {
+    if (Object.keys(envVariables).includes(extension.idEnvironmentVariableName)) {
+      extensionsIdentifiers[extension.localIdentifier] = envVariables[extension.idEnvironmentVariableName]
+    }
+  }
+  app.extensions.ui.forEach(processExtension)
+  app.extensions.function.forEach(processExtension)
+  app.extensions.theme.forEach(processExtension)
+
+  return {
+    app: envVariables[app.idEnvironmentVariableName],
+    extensions: extensionsIdentifiers,
+  }
+}
