+++ packages/app/src/cli/services/deploy/bundle.ts
@@ -1,5 +1,6 @@
 import {buildThemeExtensions, buildFunctionExtension, buildUIExtensions} from '../build/extension.js'
-import {App, Identifiers} from '../../models/app/app.js'
+import {App} from '../../models/app/app.js'
+import {Identifiers} from '../../models/app/identifiers.js'
 import {path, output, file, error} from '@shopify/cli-kit'
 import {zip} from '@shopify/cli-kit/node/archiver'
