+++ packages/app/src/cli/services/environment/id-manual-matching.ts
@@ -1,5 +1,5 @@
 import {ExtensionRegistration} from '../dev/create-extension.js'
-import {IdentifiersExtensions} from '../../models/app/app.js'
+import {IdentifiersExtensions} from '../../models/app/identifiers.js'
 import {Extension} from '../../models/app/extensions.js'
 import {ui} from '@shopify/cli-kit'
